Final Application, Executer Solver pour lancer le projet
